move mockParams
